requirements.txt
# Core dependencies
python-dotenv
together
PyPDF2
fpdf

# Optional (for future extension or Redis support)
redis

# If you're using Python 3.10+, this should work without issues.
